Rescue/requirements.txt
django==2.0.3
djangorestframework
pymongo
